+++ dpadrecyclerview/src/androidTest/kotlin/com/rubensousa/dpadrecyclerview/test/tests/layout/LayoutWhileScrollingTest.kt
@@ -65,6 +65,7 @@ class LayoutWhileScrollingTest : DpadRecyclerViewTest() {
         // given
         var layoutCompleted = 0
         onRecyclerView("Disable layout during scroll") { recyclerView ->
+            recyclerView.setLayoutWhileScrollingEnabled(false)
             recyclerView.addOnLayoutCompletedListener(
                 object : DpadRecyclerView.OnLayoutCompletedListener {
                     override fun onLayoutCompleted(state: RecyclerView.State) {