+++ dpadrecyclerview/src/main/java/com/rubensousa/dpadrecyclerview/layoutmanager/PivotLayoutManager.kt
@@ -80,13 +80,19 @@ class PivotLayoutManager(properties: Properties) : RecyclerView.LayoutManager(),
     private var isScrollingFromTouchEvent = false
     internal var layoutCompletedListener: DpadRecyclerView.OnLayoutCompletedListener? = null
 
+    // This is required for XML usage
+    @Suppress("unused", "UNUSED_PARAMETER")
+    constructor(context: Context, attrs: AttributeSet?, defStyleAttr: Int, defStyleRes: Int) : this(
+        Properties()
+    )
+
     override fun checkLayoutParams(layoutParams: RecyclerView.LayoutParams?): Boolean {
         return layoutParams is DpadLayoutParams
     }
 
     override fun generateLayoutParams(
         context: Context,
-        attrs: AttributeSet
+        attrs: AttributeSet,
     ): RecyclerView.LayoutParams {
         return DpadLayoutParams(context, attrs)
     }
@@ -139,32 +145,34 @@ class PivotLayoutManager(properties: Properties) : RecyclerView.LayoutManager(),
     override fun onLayoutChildren(recycler: RecyclerView.Recycler, state: RecyclerView.State) {
         // If we have focus, save it temporarily since the views will change and we might lose it
         hadFocusBeforeLayout = hasFocus()
-        scroller.cancelSmoothScroller()
         pivotLayout.onLayoutChildren(recycler, state)
-        layoutCompletedListener?.onLayoutCompleted(state)
+        if (hadFocusBeforeLayout) {
+            focusDispatcher.focusSelectedView()
+        }
+        if (layoutInfo.isScrollingToTarget) {
+            scroller.cancelScrollToTarget()
+        }
+        hadFocusBeforeLayout = false
     }
 
     override fun onLayoutCompleted(state: RecyclerView.State) {
+        layoutCompletedListener?.onLayoutCompleted(state)
         pivotLayout.onLayoutCompleted(state)
-        if (hadFocusBeforeLayout) {
-            focusDispatcher.focusSelectedView(recyclerView)
-        }
         pivotSelector.onLayoutCompleted()
-        hadFocusBeforeLayout = false
     }
 
     override fun collectAdjacentPrefetchPositions(
         dx: Int,
         dy: Int,
         state: RecyclerView.State,
-        layoutPrefetchRegistry: LayoutPrefetchRegistry
+        layoutPrefetchRegistry: LayoutPrefetchRegistry,
     ) {
         prefetchCollector.collectAdjacentPrefetchPositions(dx, dy, state, layoutPrefetchRegistry)
     }
 
     override fun collectInitialPrefetchPositions(
         adapterItemCount: Int,
-        layoutPrefetchRegistry: LayoutPrefetchRegistry
+        layoutPrefetchRegistry: LayoutPrefetchRegistry,
     ) {
         prefetchCollector.collectInitialPrefetchPositions(
             adapterItemCount = adapterItemCount,
@@ -177,13 +185,13 @@ class PivotLayoutManager(properties: Properties) : RecyclerView.LayoutManager(),
     override fun scrollHorizontallyBy(
         dx: Int,
         recycler: RecyclerView.Recycler,
-        state: RecyclerView.State
+        state: RecyclerView.State,
     ): Int = pivotLayout.scrollHorizontallyBy(dx, recycler, state)
 
     override fun scrollVerticallyBy(
         dy: Int,
         recycler: RecyclerView.Recycler,
-        state: RecyclerView.State
+        state: RecyclerView.State,
     ): Int = pivotLayout.scrollVerticallyBy(dy, recycler, state)
 
     override fun computeHorizontalScrollOffset(state: RecyclerView.State): Int {
@@ -278,7 +286,7 @@ class PivotLayoutManager(properties: Properties) : RecyclerView.LayoutManager(),
     override fun smoothScrollToPosition(
         recyclerView: RecyclerView,
         state: RecyclerView.State,
-        position: Int
+        position: Int,
     ) {
         scroller.scrollToPosition(position, subPosition = 0, smooth = true)
     }
@@ -291,7 +299,6 @@ class PivotLayoutManager(properties: Properties) : RecyclerView.LayoutManager(),
 
     override fun onItemsAdded(recyclerView: RecyclerView, positionStart: Int, itemCount: Int) {
         layoutInfo.invalidateSpanCache()
-        pivotLayout.onItemsAdded(positionStart, itemCount)
         pivotSelector.onItemsAdded(positionStart, itemCount)
     }
 
@@ -302,19 +309,17 @@ class PivotLayoutManager(properties: Properties) : RecyclerView.LayoutManager(),
 
     override fun onItemsRemoved(recyclerView: RecyclerView, positionStart: Int, itemCount: Int) {
         layoutInfo.invalidateSpanCache()
-        pivotLayout.onItemsRemoved(positionStart, itemCount)
         pivotSelector.onItemsRemoved(positionStart, itemCount)
     }
 
     override fun onItemsMoved(recyclerView: RecyclerView, from: Int, to: Int, itemCount: Int) {
         layoutInfo.invalidateSpanCache()
-        pivotLayout.onItemsMoved(from, to, itemCount)
         pivotSelector.onItemsMoved(from, to, itemCount)
     }
 
     override fun onAdapterChanged(
         oldAdapter: RecyclerView.Adapter<*>?,
-        newAdapter: RecyclerView.Adapter<*>?
+        newAdapter: RecyclerView.Adapter<*>?,
     ) {
         if (oldAdapter != null) {
             pivotLayout.reset()
@@ -337,14 +342,14 @@ class PivotLayoutManager(properties: Properties) : RecyclerView.LayoutManager(),
         recyclerView: RecyclerView,
         views: ArrayList<View>,
         direction: Int,
-        focusableMode: Int
+        focusableMode: Int,
     ): Boolean {
         return focusDispatcher.onAddFocusables(recyclerView, views, direction, focusableMode)
     }
 
     fun onRequestFocusInDescendants(
         direction: Int,
-        previouslyFocusedRect: Rect?
+        previouslyFocusedRect: Rect?,
     ): Boolean {
         return focusDispatcher.onRequestFocusInDescendants(direction, previouslyFocusedRect)
     }
@@ -353,7 +358,7 @@ class PivotLayoutManager(properties: Properties) : RecyclerView.LayoutManager(),
         parent: RecyclerView,
         state: RecyclerView.State,
         child: View,
-        focused: View?
+        focused: View?,
     ): Boolean {
         focusDispatcher.onRequestChildFocus(parent, child, focused)
         return true
@@ -364,7 +369,7 @@ class PivotLayoutManager(properties: Properties) : RecyclerView.LayoutManager(),
         parent: RecyclerView,
         child: View,
         rect: Rect,
-        immediate: Boolean
+        immediate: Boolean,
     ): Boolean = false
 
     override fun onAttachedToWindow(view: RecyclerView) {
@@ -386,14 +391,14 @@ class PivotLayoutManager(properties: Properties) : RecyclerView.LayoutManager(),
 
     override fun getRowCountForAccessibility(
         recycler: RecyclerView.Recycler,
-        state: RecyclerView.State
+        state: RecyclerView.State,
     ): Int {
         return accessibilityHelper.getRowCountForAccessibility(state)
     }
 
     override fun getColumnCountForAccessibility(
         recycler: RecyclerView.Recycler,
-        state: RecyclerView.State
+        state: RecyclerView.State,
     ): Int {
         return accessibilityHelper.getColumnCountForAccessibility(state)
     }
@@ -401,7 +406,7 @@ class PivotLayoutManager(properties: Properties) : RecyclerView.LayoutManager(),
     override fun onInitializeAccessibilityNodeInfo(
         recycler: RecyclerView.Recycler,
         state: RecyclerView.State,
-        info: AccessibilityNodeInfoCompat
+        info: AccessibilityNodeInfoCompat,
     ) {
         accessibilityHelper.onInitializeAccessibilityNodeInfo(recycler, state, info)
     }
@@ -410,7 +415,7 @@ class PivotLayoutManager(properties: Properties) : RecyclerView.LayoutManager(),
         recycler: RecyclerView.Recycler,
         state: RecyclerView.State,
         host: View,
-        info: AccessibilityNodeInfoCompat
+        info: AccessibilityNodeInfoCompat,
     ) {
         accessibilityHelper.onInitializeAccessibilityNodeInfoForItem(host, info)
     }
@@ -419,7 +424,7 @@ class PivotLayoutManager(properties: Properties) : RecyclerView.LayoutManager(),
         recycler: RecyclerView.Recycler,
         state: RecyclerView.State,
         action: Int,
-        args: Bundle?
+        args: Bundle?,
     ): Boolean = accessibilityHelper.performAccessibilityAction(recyclerView, state, action)
 
     override fun onSaveInstanceState(): Parcelable {
@@ -706,13 +711,13 @@ class PivotLayoutManager(properties: Properties) : RecyclerView.LayoutManager(),
     }
 
     fun addOnLayoutCompletedListener(
-        listener: DpadRecyclerView.OnLayoutCompletedListener
+        listener: DpadRecyclerView.OnLayoutCompletedListener,
     ) {
         pivotLayout.addOnLayoutCompletedListener(listener)
     }
 
     fun removeOnLayoutCompletedListener(
-        listener: DpadRecyclerView.OnLayoutCompletedListener
+        listener: DpadRecyclerView.OnLayoutCompletedListener,
     ) {
         pivotLayout.removeOnLayoutCompletedListener(listener)
     }